+++ cli/tokenize.cc
@@ -1,6 +1,7 @@
 #include <iostream>
 
 #include <boost/program_options.hpp>
+#include <boost/algorithm/string.hpp>
 
 #include <onmt/Tokenizer.h>
 
@@ -18,6 +19,8 @@ int main(int argc, char* argv[])
     ("case_feature,c", po::bool_switch()->default_value(false), "lowercase corpus and generate case feature")
     ("segment_case", po::bool_switch()->default_value(false), "Segment case feature, splits AbC to Ab C to be able to restore case")
     ("segment_numbers", po::bool_switch()->default_value(false), "Segment numbers into single digits")
+    ("segment_alphabet", po::value<std::string>()->default_value(""), "comma-separated list of alphabets on which to segment all letters.")
+    ("segment_alphabet_change", po::bool_switch()->default_value(false), "Segment if the alphabet changes between 2 letters.")
     ("bpe_model,bpe", po::value<std::string>()->default_value(""), "path to the BPE model")
     ;
 
@@ -42,18 +45,29 @@ int main(int argc, char* argv[])
     flags |= onmt::Tokenizer::Flags::SegmentCase;
   if (vm["segment_numbers"].as<bool>())
     flags |= onmt::Tokenizer::Flags::SegmentNumbers;
+  if (vm["segment_alphabet_change"].as<bool>())
+    flags |= onmt::Tokenizer::Flags::SegmentAlphabetChange;
 
-  onmt::ITokenizer* tokenizer = new onmt::Tokenizer(onmt::Tokenizer::mapMode.at(vm["mode"].as<std::string>()),
-                                                    flags,
-                                                    vm["bpe_model"].as<std::string>(),
-                                                    vm["joiner"].as<std::string>());
+  std::vector<std::string> alphabets_to_segment;
+  boost::split(alphabets_to_segment,
+               vm["segment_alphabet"].as<std::string>(),
+               boost::is_any_of(","));
+
+  onmt::Tokenizer* tokenizer = new onmt::Tokenizer(
+    onmt::Tokenizer::mapMode.at(vm["mode"].as<std::string>()),
+    flags,
+    vm["bpe_model"].as<std::string>(),
+    vm["joiner"].as<std::string>());
+
+  for (const auto& alphabet : alphabets_to_segment)
+    tokenizer->add_alphabet_to_segment(alphabet);
 
   std::string line;
 
   while (std::getline(std::cin, line))
   {
     if (!line.empty())
-      std::cout << tokenizer->tokenize(line);
+      std::cout << reinterpret_cast<onmt::ITokenizer*>(tokenizer)->tokenize(line);
 
     std::cout << std::endl;
   }