+++ include/onmt/Alphabet.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <string>
+
+#include "onmt/unicode/Unicode.h"
+
+namespace onmt
+{
+
+  std::string get_alphabet(unicode::code_point_t c);
+
+}